Fix monorepo import paths and docker-compose commands
- Updated docker-compose.yml to use pnpm commands
- Fixed relative imports in dev/ scripts (changed from @roast/web/src to relative paths)
- Fixed evaluation scripts and test data imports
- Removed duplicate next.config.ts (kept next.config.js)
