+++ .claude_workspace
@@ -1,0 +1,1 @@
+../monorepo-migration
