--- apps/web/next.config.ts
@@ -1,67 +1,0 @@
-import type { NextConfig } from "next";
-
-const securityHeaders = [
-  {
-    key: 'X-DNS-Prefetch-Control',
-    value: 'on'
-  },
-  {
-    key: 'X-Content-Type-Options',
-    value: 'nosniff'
-  },
-  {
-    key: 'X-Frame-Options',
-    value: 'DENY'
-  },
-  {
-    key: 'X-XSS-Protection',
-    value: '1; mode=block'
-  },
-  {
-    key: 'Referrer-Policy',
-    value: 'strict-origin-when-cross-origin'
-  }
-];
-
-const nextConfig: NextConfig = {
-  async headers() {
-    return [
-      {
-        source: '/:path*',
-        headers: securityHeaders,
-      },
-    ];
-  },
-  webpack: (config, { isServer }) => {
-    if (isServer) {
-      // For server-side, externalize native modules
-      config.externals = config.externals || [];
-      config.externals.push({
-        're2': 'commonjs re2',
-        'canvas': 'commonjs canvas',
-        'sharp': 'commonjs sharp',
-      });
-    } else {
-      // For client-side, completely ignore these modules
-      config.resolve.fallback = {
-        ...config.resolve.fallback,
-        're2': false,
-        'canvas': false,
-        'sharp': false,
-        'fs': false,
-        'path': false,
-        'os': false,
-      };
-    }
-    
-    // Ignore native node modules
-    config.module.rules.push({
-      test: /\.node$/,
-      use: 'ignore-loader',
-    });
-
-    return config;
-  },
-};
-
-export default nextConfig;
