+++ apps/web/src/app/api/agents/[agentId]/jobs/__tests__/route.test.ts
@@ -1,9 +1,9 @@
 import { GET } from '../route';
 import { NextRequest } from 'next/server';
-import { prisma } from '@/lib/prisma';
+import { prisma } from '@roast/db';
 
 // Mock dependencies
-jest.mock('@/lib/prisma', () => ({
+jest.mock('@roast/db', () => ({
   prisma: {
     agent: {
       findUnique: jest.fn(),
