+++ apps/web/src/app/api/user/api-keys/[keyId]/__tests__/route.test.ts
@@ -1,10 +1,10 @@
 import { DELETE } from '../route';
 import { NextRequest } from 'next/server';
-import { prisma } from '@/lib/prisma';
+import { prisma } from '@roast/db';
 import { authenticateRequestSessionFirst } from '@/lib/auth-helpers';
 
 // Mock dependencies
-jest.mock('@/lib/prisma', () => ({
+jest.mock('@roast/db', () => ({
   prisma: {
     apiKey: {
       findFirst: jest.fn(),
