+++ apps/web/src/app/api/user/api-keys/__tests__/route.test.ts
@@ -1,11 +1,11 @@
 import { GET, POST } from '../route';
 import { NextRequest } from 'next/server';
-import { prisma } from '@/lib/prisma';
+import { prisma } from '@roast/db';
 import { authenticateRequestSessionFirst } from '@/lib/auth-helpers';
 import { generateApiKey, hashApiKey } from '@/lib/crypto';
 
 // Mock dependencies
-jest.mock('@/lib/prisma', () => ({
+jest.mock('@roast/db', () => ({
   prisma: {
     apiKey: {
       findMany: jest.fn(),
