+++ apps/web/src/models/__tests__/Document.email-privacy.e2e.test.ts
@@ -1,4 +1,4 @@
-import { prisma } from '@/lib/prisma';
+import { prisma } from '@roast/db';
 import { DocumentModel } from '../Document';
 import { nanoid } from 'nanoid';
