+++ apps/web/src/models/__tests__/User.e2e.test.ts
@@ -1,4 +1,4 @@
-import { prisma } from '@/lib/prisma';
+import { prisma } from '@roast/db';
 import { UserModel } from '../User';
 
 describe('UserModel Integration Tests - Email Privacy', () => {
